+++ _torch/pyproject.toml
@@ -5,10 +5,12 @@ description = "FFTCG Image Classification and Detection."
 authors = ["Allonte <angelosheltie@gmail.com>"]
 license = "MIT License"
 readme = "../README.md"
-packages = [{include = "crystalvision", from=".."}]
+packages = [{include = "crystalvision", from="."}]
 
 [tool.poetry.dependencies]
 python = "3.11.*,<3.12"
+
+[tool.poetry.group.base.dependencies]
 pandas = ">=2.1.2"
 tqdm = ">=4.66.1"
 aiohttp = ">=3.10.2"
@@ -16,7 +18,6 @@ aiofiles = ">=23.2.1"
 dogpile-cache = ">=1.3.1"
 # gradio = "^4.39.0"
 swifter = {extras = ["groupby"], version = ">=1.4.0"}
-# playwright = "^1.47.0"
 
 [tool.poetry.group.ml.dependencies]
 keras = ">3.4.0"
@@ -86,6 +87,7 @@ pypdf = ">=5.0.1"
 playwright = ">=1.47.0"
 jq = ">=1.8.0"
 Jinja2 = ">=3.1.4"
+langchain-experimental = ">=0.3.2"
 
 [tool.poetry.group.discord.dependencies]
 discord-py = ">=2.4.0"
