+++ crystalvision/lang/df_description.json
@@ -1,0 +1,26 @@
+{
+    "code": {
+        "description": "the UUID of the card"
+    },
+    "name_en": {
+        "description": "the English name of the card"
+    },
+    "type_en": {
+        "description": "the English type of the card"
+    },
+    "isMultiElement": {
+        "description": "if the card is multielement/multi-element/multi element but not which elements it is"
+    },
+    "element": {
+        "description": "each element of the card as python set"
+    },
+    "element_ja": {
+        "description": "each element of the card in Japanese as python set"
+    },
+    "numElements": {
+        "description": "the number of elements the card consists of"
+    },
+    "ex_burst": {
+        "description": "if the card has an exburst/ex burst/ex-burst/ex/EX ability"
+    }
+}
